+++ NEWS.md
@@ -1,3 +1,13 @@
+# maybe 1.0.0
+
+Features
+
+- `maybe_map2` Evaluate a binary function on two maybe values
+- `maybe_case` Unwrap and call a function on a maybe value or return a default
+- `from_just` Unwrap a 'Just' value or throw an error
+- `filter_justs` Filter and unwrap a list of 'Just' values
+- `filter_map` Map a function over a list and filter only 'Just' values
+
 # maybe 0.2.1
 
 Internal changes